+++ CHANGELOG.md
@@ -2,7 +2,8 @@ QuantumCircuitOpt.jl Change Log
 ===============================
 
 ### v0.5.5
-- `optimal_global_phase` now recognizes commuting elemetary gate pairs that commute up to a phase
+- `optimal_global_phase` now recognizes commuting elementary gate pairs that commute up to a global phase
+- Docs updated with the SC22 (IEEE) paper link
 - Updated docs and unit tests to reflect above changes
 
 ### v0.5.4