+++ CHANGELOG.md
@@ -6,6 +6,7 @@ QuantumCircuitOpt.jl Change Log
 - Docs updated with the SC22 (IEEE) paper link
 - Added `isapprox_global_phase` utility to evaluate equivalences of any two complex gates up to a global phase
 - Updated docs and unit tests to reflect above changes
+- Removed compatability on Pkg
 
 ### v0.5.4
 - Added a generalized function to obtain controlled gates (`controlled_gate()`) in any number of qubits